[dagit] Fix right Asset Graph panel missing if saved "explorer" state is absent
